Move diff derivation to one spot
